pyproject.toml
[project]
name = "mwet-schemas"
version = "0.0.1"
authors = [
  { name="Runbo Jiang", email="rjiang2@lbl.gov" },
]
description = "A package to ingest nmr data to SciCat"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
          "linkml"
          ]

[project.optional-dependencies]
ingest = [
  "python-dotenv",
  "pandas",
  "pyscicat",
  "typer",
  "matplotlib",
]

# These are required for developing the package (running the tests, building
# the documentation) but not necessarily required for _using_ it.
dev = [
    "flake8",
    "pre-commit",
    "pytest-mock"
]

[project.urls]
Homepage = "https://github.com/als-computing/mwet-schemas"
Issues = "https://github.com/als-computing/mwet-schemas/issues"


[tool.isort]
profile = "black"
